# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: My Jeja API
  version: 1.1.0
  description: |
    MVP API for My Jeja.
    Scope: math-only, levels 1-9, streaming teaching chat, secure exam grading.
servers:
  - url: https://your-service.up.railway.app/api/v1
...
paths:
...
  /subjects:
    get:
      tags: [Dashboard]
      summary: List subjects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subject list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectOut'
    post:
      tags: [Dashboard]
      summary: Create subject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreateRequest'
      responses:
        '200':
          description: Subject created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectOut'
  /subjects/{subject_id}/curriculum:
    get:
      tags: [Dashboard]
      summary: List subject curriculum
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subject_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Curriculum list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CurriculumOut'
    post:
      tags: [Dashboard]
      summary: Create curriculum item
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subject_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurriculumCreateRequest'
      responses:
        '200':
          description: Curriculum item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumOut'
  /subjects/{subject_id}/stages:
    get:
      tags: [Dashboard]
      summary: List stages with exam unlock status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subject_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stage list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StageOut'
    post:
      tags: [Dashboard]
      summary: Create stage
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subject_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StageCreateRequest'
      responses:
        '200':
          description: Stage created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageOut'
  /subjects/{subject_id}/stages/{stage_id}/exams:
    post:
      tags: [Exam]
      summary: Create stage exam (unlocked only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: subject_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: stage_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exam created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateExamResponse'
        '422':
          description: Exam locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SubjectCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
    SubjectOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
    CurriculumCreateRequest:
      type: object
      required: [title]
      properties:
        title: { type: string }
        note: { type: string, nullable: true }
        order_index: { type: integer, nullable: true }
    CurriculumOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        subject_id: { type: string, format: uuid }
        title: { type: string }
        note: { type: string, nullable: true }
        order_index: { type: integer }
    StageCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        order_index: { type: integer, nullable: true }
        curriculum_item_ids:
          type: array
          items: { type: string, format: uuid }
    StageOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        subject_id: { type: string, format: uuid }
        name: { type: string }
        order_index: { type: integer }
        passed: { type: boolean }
        exam_unlocked: { type: boolean }
        untaught_count: { type: integer }
...
    ExamQuestion:
      type: object
      properties:
        question_id: { type: string, format: uuid }
        question_no: { type: integer }
        type: { type: string, enum: [multiple_choice, short_answer] }
        content: { type: string }
        options:
          type: array
          items: { type: string }
          nullable: true
    CreateExamResponse:
      type: object
      properties:
        exam_id: { type: string, format: uuid }
        level: { type: integer, minimum: 1, maximum: 9 }
        questions:
          type: array
          items:
            $ref: '#/components/schemas/ExamQuestion'
...
    ErrorResponse:
      type: object
      properties:
        detail: { type: string }
        request_id: { type: string }